upload fix

+++ toh-pt6/src/app/users/users.component.ts
@@ -12,13 +12,20 @@ export class UsersComponent implements OnInit {
 
   users: User[]  = [];
   roles: Role[] = [];
-  selectedRole!: Role;
+  //selectedRole!: Role;
+  selectedRole: any = null;
 
   constructor(private roleService: RoleService,private userService: UserService) { }
 
   ngOnInit() {
     this.getUsers();
     this.getRoles();
+    this.selectedRole = this.roles[0]
+    /*
+    if (this.roles.length > 0) {
+      this.selectedRole = this.roles[0];
+    }
+      */
 
   }
 
@@ -29,7 +36,7 @@ export class UsersComponent implements OnInit {
 
   getRoles(): void {
     this.roleService.getRoles().subscribe(roles => {
-      console.log('Roles:', roles);
+      //console.log('Roles:', roles);
       this.roles = roles;
     });
   }
@@ -68,22 +75,29 @@ export class UsersComponent implements OnInit {
     if (user.id === undefined) {
       console.error('User ID is undefined. Cannot delete user.');
       return;
-  }
-
-  // Odstranění uživatele z lokálního seznamu
-  this.users = this.users.filter(u => u.id !== user.id);
-
-  // Volání API pro odstranění uživatele
-  this.userService.deleteUser(user.id).subscribe({
+    }
+  
+    // Zobrazení potvrzovacího dialogu
+    const confirmDelete = confirm('Jste si jisti, že chcete uživatele odstranit?');
+    if (!confirmDelete) {
+      return; // Pokud uživatel klikne na "Ne", operace se přeruší
+    }
+  
+    // Odstranění uživatele z lokálního seznamu
+    this.users = this.users.filter(u => u.id !== user.id);
+  
+    // Volání API pro odstranění uživatele
+    this.userService.deleteUser(user.id).subscribe({
       next: () => {
-          console.log('User deleted successfully.');
+        console.log('User deleted successfully.');
       },
       error: (err) => {
-          console.error('Failed to delete user:', err);
-          // Můžete přidat kód pro obnovu seznamu uživatelů, pokud odstranění selže
-          // this.users.push(user); // např. znovu přidat uživatele do seznamu
+        console.error('Failed to delete user:', err);
+        // Můžete přidat kód pro obnovu seznamu uživatelů, pokud odstranění selže
+        // this.users.push(user); // např. znovu přidat uživatele do seznamu
       }
-  });
+    });
   }
+  
 }
